[UBS] - The cost of the service and tariff service, as well as the order cost, cannot be saved to the nearest kopeck #5935 (#1134)

+++ service-api/src/main/java/greencity/dto/order/OrderStatusPageDto.java
@@ -39,7 +39,7 @@ public class OrderStatusPageDto {
     private ExportDetailsDto exportDetailsDto;
     private EmployeePositionDtoRequest employeePositionDtoRequest;
     private String comment;
-    private Integer courierPricePerPackage;
+    private Double courierPricePerPackage;
     private CourierInfoDto courierInfo;
-    private Long writeOffStationSum;
+    private Double writeOffStationSum;
 }
